// Inzynierka/frontend/src/app/shared/models/messages/showAuctionMessages.ts
/**
 * Created by DZONI on 07.01.2017.
 */
export interface ShowAuctionMessages {
    countryFrom: string;
    countryTo: string;
    deliveryDate: string;
    description: string;
    endOfAuction: string;
    ended: string;
    footer: string;
    fragile: string;
    fullAddressFrom: string;
    fullAddressTo: string;
    height: string;
    id: string;
    length: string;
    living: string;
    loadDate: string;
    premium: string;
    specialEnvironment: string;
    startOfAuctionDate: string;
    title: string;
    weight: string;
    width: string;
}